Переносит логику по рендерингу списков фильмов в PageController
